feat(#14): move classes into separate packages
