+++ src/main/java/com/github/lombrozo/testnames/rules/PresentTense.java
@@ -22,42 +22,60 @@
  * SOFTWARE.
  */
 
-package com.github.lombrozo.testnames;
+package com.github.lombrozo.testnames.rules;
 
-import java.util.Arrays;
-import java.util.Collection;
+import com.github.lombrozo.testnames.Rule;
+import com.github.lombrozo.testnames.TestCase;
+import com.github.lombrozo.testnames.WrongTestName;
 
 /**
- * The rule checks if test case in present simple.
+ * The rule checks if test case in present tense.
  *
  * @since 0.1.0
  */
-public final class PresentSimpleRule implements Rule {
+public final class PresentTense implements Rule {
 
     /**
-     * The rules.
+     * The test case.
      */
-    private final Collection<Rule> all;
+    private final TestCase test;
 
     /**
      * Ctor.
      *
      * @param test The test case to check
      */
-    public PresentSimpleRule(final TestCase test) {
-        this.all = Arrays.asList(
-            new NotCamelCase(test),
-            new NotContainsTestWord(test),
-            new NotSpam(test),
-            new NotUsesSpecialCharacters(test),
-            new PresentTense(test)
-        );
+    public PresentTense(final TestCase test) {
+        this.test = test;
     }
 
     @Override
     public void validate() throws WrongTestName {
-        for (final Rule rule : this.all) {
-            rule.validate();
+        if (!this.presentTense()) {
+            throw new WrongTestName(
+                this.test,
+                "the test name has to be written using present tense"
+            );
         }
+    }
+
+    /**
+     * Is test case name in present tense.
+     *
+     * @return The result
+     * @checkstyle ReturnCountCheck (20 lines)
+     */
+    @SuppressWarnings("PMD.OnlyOneReturn")
+    private boolean presentTense() {
+        final char[] chars = this.test.name().toCharArray();
+        char prev = '!';
+        for (final char chr : chars) {
+            if (Character.isUpperCase(chr)) {
+                return prev == 's';
+            } else {
+                prev = chr;
+            }
+        }
+        return prev == 's';
     }
 }
