+++ src/test/java/com/github/lombrozo/testnames/rules/RuleTest.java
@@ -22,14 +22,16 @@
  * SOFTWARE.
  */
 
-package com.github.lombrozo.testnames;
+package com.github.lombrozo.testnames.rules;
 
+import com.github.lombrozo.testnames.TestCase;
+import com.github.lombrozo.testnames.WrongTestName;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 /**
- * Test case for {@link Rule}.
+ * Test case for {@link com.github.lombrozo.testnames.Rule}.
  *
  * @since 0.1.0
  */
